GA-44 separated public and private views
